test-util: extract shared RamDisk helper for ext2 integration tests
Closes #627.

Every ext2 integration test crate carries a near-identical ~70-line
copy of the same `RamDisk` struct: a `Mutex<Vec<u8>>`-backed
`BlockDevice` impl with a write-count atomic and a read-only latch.
CodeRabbit flagged the duplication on #626. This change extracts the
helper to `kernel/tests/common/ext2_ramdisk.rs` and switches the
three call sites named in the issue (`ext2_ialloc.rs`,
`ext2_dir_ops.rs`, `ext2_unlink.rs`) to use it.

The helper is included via the per-file

    #[path = "common/ext2_ramdisk.rs"]
    mod ext2_ramdisk;
    use ext2_ramdisk::RamDisk;

pattern. Each integration test under `kernel/tests/` is its own
`#![no_std] #![no_main]` crate, so a normal `mod common` reference
would not resolve and would also try to compile the file as its own
test binary. The `#[path = ...]` include keeps the file an inline
module of the consumer crate, which compiles cleanly under the
no_std / no_main harness without ever being built standalone.
The file lives in `tests/common/` (a subdirectory) so cargo's
autotest discovery skips it; no `[[test]]` entry is needed.

Surface is intentionally minimal: `from_image`, `set_read_only`,
`writes`, plus the `BlockDevice` impl. `writes()` is included
because `ext2_dir_ops.rs` consumes it for the
"RO mount must not issue any writes" assertions; without it that
call site would lose coverage.

Behaviour is unchanged. The unified write_at gates on the `read_only`
latch unconditionally — that gate already existed in the
`ext2_dir_ops.rs` and `ext2_unlink.rs` copies; for `ext2_ialloc.rs`
the latch defaults to `false` so the path stays a no-op there.

Net diff: -202 LoC across the three test files.
